Fix two uncaught exceptions
There are two uncaught exceptions that the nrjmx tool is not
currently catching that were causing cryptic failures in the
newrelic-integrations-sdk jmx library. One was
RuntimeMbeanException, the other RuntimeOperationsException,
both caused by trying to collect unavailable attributes.

This commit adds both exceptions to the list of errors that
are logged and ignored so collection can continue.