Add fix testerver to work with modules
The testerver project will put all of the appropriate files in
a clocation so it can be built with Docker. Unlike the Maven build,
the Gradle build will not directly build the container, but the
test code will invoke thew building of the container by reading
a system property for the location of the files.

Due to an unfotunate dependency on JUnit4 code inside the 1.x
series of testcontainers, using JUnit5 + modules requires a
workaround. For this reason the fake-junt4-dependencies subproject
is required. It adds a single RestRule class which is exported as a
project and which is then used during testing.